Evaluate JSON in new Roku pkg web interface
The old regexes to capture the result and the package location weren't
working.  I was able to refine the regex for the result, but upon
further investigation Roku seems to have changed the interface to insert
embedded JSON data dynamically into the page with JS, so the package
regex isn't working at all.  Update the script so it will try to get the
required package location from the JSON, but will still fall back to the
old regex if that fails, in case someone is working on a Roku that
hasn't been updated.